Use path/filepath over hardcoded path separators.
What
===
Change tests to use `path/filepath` instead of `fmt` and hardcoded `/`
path separators.

Why
===
Joining paths with `/` causes the tests to fail when run on windows because it's file path separator is `\`. Using `path/filepath` allows the code to join paths together using the system path separator without the code needing to know what the separator is.