Allow paths starting with a tilde
Path expansion currently disallows paths starting with `~`.
The `~` character is a valid part of a directory or file name.
A leading `~` is most commonly found in university server paths
where the users have a public web directory, e.g., https://cs.fit.edu/~mmahoney/.

Return the path as-is if the leading `~` is not referring to a
directory name as in `~/` or `~`.